+++ src/controllers/user.controller.ts
@@ -95,7 +95,7 @@ const deleteUser = async (req: Request, res: Response) => {
   try {
     const filename = await getUserImage(user_id);
     await prisma.users.delete({ where: { user_id: user_id } })
-    fs.unlinkSync(path.join('public/images/user', String(filename?.split("/")[3])));
+    if (req.file !== undefined) fs.unlinkSync(path.join('public/images/user', String(filename?.split("/")[3])));
     res.json({ status: 200, msg: 'Delete Successful!!' });
   } catch (err) {
     res.status(400).json({ error: err });
